refactor: s3 패키지 구성 수정
