+++ src/main/java/org/ahpuh/surf/post/service/PostService.java
@@ -17,8 +17,8 @@ import org.ahpuh.surf.post.domain.repository.PostRepository;
 import org.ahpuh.surf.post.dto.PostScoreCategoryDto;
 import org.ahpuh.surf.post.dto.request.PostRequestDto;
 import org.ahpuh.surf.post.dto.response.*;
-import org.ahpuh.surf.s3.FileStatus;
-import org.ahpuh.surf.s3.S3Service;
+import org.ahpuh.surf.s3.domain.FileStatus;
+import org.ahpuh.surf.s3.service.S3Service;
 import org.ahpuh.surf.user.domain.User;
 import org.ahpuh.surf.user.domain.UserRepository;
 import org.springframework.data.domain.PageRequest;
