--- src/main/java/org/ahpuh/surf/s3/FileStatus.java
@@ -1,7 +1,0 @@
-package org.ahpuh.surf.s3;
-
-public record FileStatus(
-        String fileUrl,
-        FileType fileType
-) {
-}
