--- src/main/java/org/ahpuh/surf/s3/FileType.java
@@ -1,5 +1,0 @@
-package org.ahpuh.surf.s3;
-
-public enum FileType {
-    IMG, FILE
-}
