+++ src/main/java/org/ahpuh/surf/s3/domain/FileStatus.java
@@ -1,0 +1,6 @@
+package org.ahpuh.surf.s3.domain;
+
+public record FileStatus(
+        String fileUrl,
+        FileType fileType) {
+}
