+++ src/main/java/org/ahpuh/surf/s3/domain/FileType.java
@@ -1,0 +1,5 @@
+package org.ahpuh.surf.s3.domain;
+
+public enum FileType {
+    IMAGE, FILE
+}
