+++ src/test/java/org/ahpuh/surf/unit/user/service/UserServiceTest.java
@@ -3,7 +3,7 @@ package org.ahpuh.surf.unit.user.service;
 import org.ahpuh.surf.common.exception.user.DuplicatedEmailException;
 import org.ahpuh.surf.common.exception.user.InvalidPasswordException;
 import org.ahpuh.surf.common.exception.user.UserNotFoundException;
-import org.ahpuh.surf.s3.S3Service;
+import org.ahpuh.surf.s3.service.S3Service;
 import org.ahpuh.surf.user.domain.User;
 import org.ahpuh.surf.user.domain.UserConverter;
 import org.ahpuh.surf.user.domain.UserRepository;
